Refactor session cookie definition to common session package
Session cookie is now used by both IdP sessions and OAuth 'sessions'
