+++ pkg/lib/deps/deps_utils.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/interaction"
+	"github.com/authgear/authgear-server/pkg/lib/session"
 	"github.com/authgear/authgear-server/pkg/lib/session/idpsession"
 	"github.com/authgear/authgear-server/pkg/util/httputil"
 )
@@ -15,6 +16,7 @@ var utilsDeps = wire.NewSet(
 	wire.NewSet(
 		httputil.DependencySet,
 		NewCookieFactory,
+		wire.Bind(new(session.CookieFactory), new(*httputil.CookieFactory)),
 		wire.Bind(new(idpsession.CookieFactory), new(*httputil.CookieFactory)),
 		wire.Bind(new(interaction.CookieFactory), new(*httputil.CookieFactory)),
 	),
